moved helper functions to utils
