+++ components/icons/future_threshold.tsx
@@ -15,8 +15,8 @@ export const FutureThresholdLegendIcon: FC = ({}) => (
     <path
       d="M73 5C67.9282 5 63.8011 5 59.7238 5C55.6464 5 56.0689 5 51.4199 5C50.2247 5 46 5 46 5"
       stroke="#E0E0E0"
-      stroke-width="2"
-      stroke-dasharray="2 2"
+      strokeWidth="2"
+      strokeDasharray="2 2"
     />
     <path
       d="M0.574929 12.9375V12.1818H5.07777V12.9375H3.26243V18H2.38743V12.9375H0.574929ZM6.85405 15.4091V18H6.00462V12.1818H6.84268V14.3466H6.89666C6.99893 14.1117 7.15518 13.9252 7.36541 13.7869C7.57564 13.6487 7.85026 13.5795 8.18928 13.5795C8.48852 13.5795 8.74988 13.6411 8.97337 13.7642C9.19875 13.8873 9.37299 14.071 9.49609 14.3153C9.62109 14.5578 9.68359 14.8608 9.68359 15.2244V18H8.83416V15.3267C8.83416 15.0066 8.75178 14.7585 8.587 14.5824C8.42223 14.4044 8.19306 14.3153 7.8995 14.3153C7.69875 14.3153 7.51882 14.358 7.35973 14.4432C7.20253 14.5284 7.07848 14.6534 6.98757 14.8182C6.89856 14.9811 6.85405 15.178 6.85405 15.4091ZM10.8171 18V13.6364H11.6381V14.3295H11.6836C11.7631 14.0947 11.9033 13.91 12.104 13.7756C12.3067 13.6392 12.5359 13.571 12.7915 13.571C12.8446 13.571 12.9071 13.5729 12.979 13.5767C13.0529 13.5805 13.1107 13.5852 13.1523 13.5909V14.4034C13.1183 14.3939 13.0576 14.3835 12.9705 14.3722C12.8834 14.3589 12.7963 14.3523 12.7092 14.3523C12.5084 14.3523 12.3294 14.3949 12.1722 14.4801C12.0169 14.5634 11.8938 14.6799 11.8029 14.8295C11.712 14.9773 11.6665 15.1458 11.6665 15.3352V18H10.8171ZM15.6119 18.0881C15.1819 18.0881 14.8117 17.9962 14.5011 17.8125C14.1924 17.6269 13.9537 17.3665 13.7852 17.0312C13.6185 16.6941 13.5352 16.2992 13.5352 15.8466C13.5352 15.3996 13.6185 15.0057 13.7852 14.6648C13.9537 14.3239 14.1886 14.0578 14.4897 13.8665C14.7927 13.6752 15.1469 13.5795 15.5522 13.5795C15.7984 13.5795 16.0371 13.6203 16.2681 13.7017C16.4992 13.7831 16.7066 13.911 16.8903 14.0852C17.074 14.2595 17.2189 14.4858 17.3249 14.7642C17.431 15.0407 17.484 15.3769 17.484 15.7727V16.0739H14.0153V15.4375H16.6516C16.6516 15.214 16.6062 15.0161 16.5153 14.8438C16.4244 14.6695 16.2965 14.5322 16.1317 14.4318C15.9689 14.3314 15.7776 14.2812 15.5579 14.2812C15.3192 14.2812 15.1109 14.34 14.9329 14.4574C14.7567 14.5729 14.6204 14.7244 14.5238 14.9119C14.4291 15.0975 14.3817 15.2992 14.3817 15.517V16.0142C14.3817 16.3059 14.4329 16.554 14.5352 16.7585C14.6393 16.9631 14.7842 17.1193 14.9698 17.2273C15.1554 17.3333 15.3723 17.3864 15.6204 17.3864C15.7814 17.3864 15.9281 17.3636 16.0607 17.3182C16.1933 17.2708 16.3079 17.2008 16.4045 17.108C16.5011 17.0152 16.5749 16.9006 16.6261 16.7642L17.43 16.9091C17.3656 17.1458 17.2501 17.3532 17.0835 17.5312C16.9187 17.7074 16.7113 17.8447 16.4613 17.9432C16.2132 18.0398 15.93 18.0881 15.6119 18.0881ZM21.6964 14.7017L20.9265 14.8381C20.8943 14.7396 20.8432 14.6458 20.7731 14.5568C20.7049 14.4678 20.6121 14.3949 20.4947 14.3381C20.3772 14.2812 20.2305 14.2528 20.0543 14.2528C19.8138 14.2528 19.613 14.3068 19.4521 14.4148C19.2911 14.5208 19.2106 14.6581 19.2106 14.8267C19.2106 14.9725 19.2646 15.09 19.3725 15.179C19.4805 15.268 19.6547 15.3409 19.8952 15.3977L20.5884 15.5568C20.9899 15.6496 21.2892 15.7926 21.4862 15.9858C21.6831 16.179 21.7816 16.4299 21.7816 16.7386C21.7816 17 21.7058 17.233 21.5543 17.4375C21.4047 17.6402 21.1954 17.7992 20.9265 17.9148C20.6594 18.0303 20.3498 18.0881 19.9975 18.0881C19.5089 18.0881 19.1102 17.9839 18.8015 17.7756C18.4928 17.5653 18.3034 17.267 18.2333 16.8807L19.0543 16.7557C19.1055 16.9697 19.2106 17.1316 19.3697 17.2415C19.5288 17.3494 19.7362 17.4034 19.9918 17.4034C20.2702 17.4034 20.4928 17.3456 20.6594 17.2301C20.8261 17.1127 20.9094 16.9697 20.9094 16.8011C20.9094 16.6648 20.8583 16.5502 20.756 16.4574C20.6557 16.3646 20.5013 16.2945 20.293 16.2472L19.5543 16.0852C19.1471 15.9924 18.846 15.8447 18.6509 15.642C18.4577 15.4394 18.3612 15.1828 18.3612 14.8722C18.3612 14.6146 18.4331 14.3892 18.5771 14.196C18.721 14.0028 18.9199 13.8523 19.1737 13.7443C19.4274 13.6345 19.7182 13.5795 20.0458 13.5795C20.5174 13.5795 20.8886 13.6818 21.1594 13.8864C21.4303 14.089 21.6093 14.3608 21.6964 14.7017ZM23.565 15.4091V18H22.7156V12.1818H23.5536V14.3466H23.6076C23.7099 14.1117 23.8661 13.9252 24.0763 13.7869C24.2866 13.6487 24.5612 13.5795 24.9002 13.5795C25.1995 13.5795 25.4608 13.6411 25.6843 13.7642C25.9097 13.8873 26.0839 14.071 26.207 14.3153C26.332 14.5578 26.3945 14.8608 26.3945 15.2244V18H25.5451V15.3267C25.5451 15.0066 25.4627 14.7585 25.2979 14.5824C25.1332 14.4044 24.904 14.3153 24.6104 14.3153C24.4097 14.3153 24.2298 14.358 24.0707 14.4432C23.9135 14.5284 23.7894 14.6534 23.6985 14.8182C23.6095 14.9811 23.565 15.178 23.565 15.4091ZM29.3661 18.0881C28.957 18.0881 28.6 17.9943 28.2951 17.8068C27.9902 17.6193 27.7534 17.357 27.5849 17.0199C27.4163 16.6828 27.332 16.2888 27.332 15.8381C27.332 15.3854 27.4163 14.9896 27.5849 14.6506C27.7534 14.3116 27.9902 14.0483 28.2951 13.8608C28.6 13.6733 28.957 13.5795 29.3661 13.5795C29.7752 13.5795 30.1322 13.6733 30.4371 13.8608C30.7421 14.0483 30.9788 14.3116 31.1474 14.6506C31.3159 14.9896 31.4002 15.3854 31.4002 15.8381C31.4002 16.2888 31.3159 16.6828 31.1474 17.0199C30.9788 17.357 30.7421 17.6193 30.4371 17.8068C30.1322 17.9943 29.7752 18.0881 29.3661 18.0881ZM29.369 17.375C29.6341 17.375 29.8538 17.3049 30.0281 17.1648C30.2023 17.0246 30.3311 16.8381 30.4144 16.6051C30.4996 16.3722 30.5423 16.1155 30.5423 15.8352C30.5423 15.5568 30.4996 15.3011 30.4144 15.0682C30.3311 14.8333 30.2023 14.6449 30.0281 14.5028C29.8538 14.3608 29.6341 14.2898 29.369 14.2898C29.1019 14.2898 28.8803 14.3608 28.7042 14.5028C28.5299 14.6449 28.4002 14.8333 28.315 15.0682C28.2317 15.3011 28.19 15.5568 28.19 15.8352C28.19 16.1155 28.2317 16.3722 28.315 16.6051C28.4002 16.8381 28.5299 17.0246 28.7042 17.1648C28.8803 17.3049 29.1019 17.375 29.369 17.375ZM33.1978 12.1818V18H32.3484V12.1818H33.1978ZM35.9741 18.0852C35.6218 18.0852 35.3074 17.9953 35.0309 17.8153C34.7563 17.6335 34.5404 17.375 34.3832 17.0398C34.2279 16.7027 34.1502 16.2983 34.1502 15.8267C34.1502 15.3551 34.2288 14.9517 34.386 14.6165C34.5451 14.2812 34.7629 14.0246 35.0394 13.8466C35.3159 13.6686 35.6294 13.5795 35.9798 13.5795C36.2506 13.5795 36.4684 13.625 36.6332 13.7159C36.7998 13.8049 36.9286 13.9091 37.0195 14.0284C37.1123 14.1477 37.1843 14.2528 37.2354 14.3438H37.2866V12.1818H38.136V18H37.3065V17.321H37.2354C37.1843 17.4138 37.1104 17.5199 37.0138 17.6392C36.9192 17.7585 36.7885 17.8627 36.6218 17.9517C36.4551 18.0407 36.2392 18.0852 35.9741 18.0852ZM36.1616 17.3608C36.4059 17.3608 36.6123 17.2964 36.7809 17.1676C36.9513 17.0369 37.0801 16.8561 37.1673 16.625C37.2563 16.3939 37.3008 16.125 37.3008 15.8182C37.3008 15.5152 37.2572 15.25 37.1701 15.0227C37.083 14.7955 36.9551 14.6184 36.7866 14.4915C36.618 14.3646 36.4097 14.3011 36.1616 14.3011C35.9059 14.3011 35.6928 14.3674 35.5224 14.5C35.3519 14.6326 35.2231 14.8134 35.136 15.0426C35.0508 15.2718 35.0082 15.5303 35.0082 15.8182C35.0082 16.1098 35.0517 16.3722 35.1388 16.6051C35.226 16.8381 35.3548 17.0227 35.5252 17.1591C35.6976 17.2936 35.9097 17.3608 36.1616 17.3608Z"
@@ -25,8 +25,8 @@ export const FutureThresholdLegendIcon: FC = ({}) => (
     <path
       d="M73 15C67.9282 15 63.8011 15 59.7238 15C55.6464 15 56.0689 15 51.4199 15C50.2247 15 46 15 46 15"
       stroke="#FFD053"
-      stroke-width="2"
-      stroke-dasharray="2 2"
+      strokeWidth="2"
+      strokeDasharray="2 2"
     />
   </svg>
 );
