+++ pages/dashboard.tsx
@@ -1,7 +1,7 @@
 import NumberDisplay from 'components/NumberDisplay';
 import Warnings from 'components/Warnings';
 import Graph from 'components/Graph';
-import FutureGraph from 'components/FutureGraph';
+import ThresholdGraph from '~/components/ThresholdGraph';
 import { withPageAuthRequired } from '@auth0/nextjs-auth0';
 
 const Dashboard = () => {
@@ -9,10 +9,7 @@ const Dashboard = () => {
     <div className="bg-ocf-black w-screen h-screen px-4">
       <h1 className="mt-4 text-ocf-gray text-3xl font-bold">Dashboard</h1>
       <div className="flex flex-row w-full justify-start">
-        <Warnings />
-      </div>
-      <div className="flex flex-row w-full justify-start">
-        <FutureGraph />
+        <ThresholdGraph />
       </div>
       <div className="flex flex-row w-full justify-start">
         <NumberDisplay title="Today's Expected Output" value="2700kW" />
